generated_projects/flask-recipe-finder_2ccf7c7d/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5
passlib==1.7.4
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
spoonacular==3.0
Flask-Caching==2.1.0
Pillow==10.1.0